Validation ID Number
The validity of the customer ID Number is very necessary to ensure that the data we take is correct. Based on these results, there are differences in the number of ID numbers from the first data loaded to the final result. The number of rows of data when it is first loaded is 7113 rows and 22 columns with 7017 unique IDs. Then after checking the validity of the customer ID, the remaining 6993 rows of data